+++ packages/component/button/component.js
@@ -1,8 +1,8 @@
 import { html } from '@dom'
 
-function component() {
+function component(button) {
   return html`
-    <button>
+    <button ${button.alt ? `aria-label="${button.alt}"` : ''}>
       <slot></slot>
     </button>
   `